monitoramento_ambiental/serializers/propriedade_serializer.py
from rest_framework import serializers
from ..models.propriedade import Propriedade

class PropriedadeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Propriedade
        fields = '__all__'